GO2heatmap {annotate}	R Documentation

Compute a heatmap for the specified data, for either a GO category or a KEGG pathway.

Description

For a given GO category or KEGG pathway, all probes in the supplied data are mapped to the pathway and a heatmap is produced.

Usage


GO2heatmap(x, eset, data, ...)
KEGG2heatmap(x, eset, data, ...)

Arguments

x
The name of the category or pathway.

eset
An ExpressionSet providing the data.

data
The name of the chip.

...
Additional parameters to pass to heatmap.

Details

For the given pathway or GO category all matching probes are determined, these are used to subset the data and heatmap is invoked on that set of data. Extra parameters can be passed through to heatmap using the ... parameter. The annotation slot of the eset argument is used to determine the appropriate annotation data to use.

Value

The value returned by heatmap is passed back to the user.

Author(s)

R. Gentleman

See Also

heatmap

Examples


  library("hgu95av2.db")
  data(sample.ExpressionSet)
  KEGG2heatmap("04810", sample.ExpressionSet, "hgu95av2")

[Package annotate version 1.60.1 Index]