GOmnplot {annotate}	R Documentation

A function to plot by group means against each other.

Description

For a two sample comparison, as determined by group, and a specified KEGG pathway or GO category, per group means are computed and plotted against each other.

Usage


GOmnplot(x, eset, data = "hgu133plus2", group, ...)
KEGGmnplot(x, eset, data = "hgu133plus2", group, ...)

Arguments

x
The name of the KEGG pathway or GO category.

eset
An ExpressionSet containing the data.

data
The name of the chip that was used to provide the data.

group
The variable indicating group membership, should have two different values.

...
Extra parameters to pass to the call to plot.

Details

All probes in eset that map to the given category are determined. Then per group, per probe means are computed and plotted against each other. Extra parameters can be passed to the plot function via the dots argument.

Value

The matrix of per group means, for each probe.

Author(s)

R. Gentleman

See Also

KEGG2heatmap

Examples


  library("hgu95av2.db")
  data(sample.ExpressionSet)
  KEGGmnplot("04810", sample.ExpressionSet, sample.ExpressionSet$sex, 
             data = "hgu95av2")

[Package annotate version 1.60.1 Index]