HTMLPage-class {annotate}	R Documentation

Classes to represent HTML pages

Description

Class HTMLPage and FramedHTMLPage are a pair of experimental classes used to explore concepts of representing HTML pages using S4 objects.

Slots

fileName:

Object of class "character" The filename of the HTML page

pageText:

Object of class "character" The text of the HTML page

pageTitle:

Object of class "character" The title of the HTML page

topPage:

Object of class "HTMLPage" The header page for a FramedHTMLPage

sidePage:

Object of class "HTMLPage" The side index page for a FramedHTMLPage

mainPage:

Object of class "HTMLPage" The primary page for a FramedHTMLPage

Methods

show

signature(object = "HTMLPage"): Describes information about the page

fileName

signature(object = "HTMLPage"): Gets the fileName slot

pageText

signature(object = "HTMLPage"): Gets the pageText slot

pageTitle

signature(object = "HTMLPage"): Gets the pageTitle slot

toFile

signature(object = "HTMLPage"): Writes the page out to the file designated by the fileName slot

Note

These classes are currently experimental.

FramedHTMLPage is modeled after the framing layout of the Bioconductor website (www.bioconductor.org).

Author(s)

Jeff Gentry

Examples


##---- Should be DIRECTLY executable !! ----

[Package annotate version 1.60.1 Index]