UniGeneQuery {annotate}	R Documentation

Create a Query String for a UniGene Identifier

Description

Given a set of UniGene identifiers this function creates a set of URLs that an be used to either open a browser to the requested location or that can be used as anchors in the construction of HTML output.

Usage


UniGeneQuery(query, UGaddress="UniGene/", type="CID")

Arguments

query
The UniGene identifiers.

UGaddress
The address of UniGene, within the NCBI repository.

type
What type of object is being asked for; eithe CID or UGID

Details

Using published details from NCBI we construct an appropriate string for directing a web browser to the information available at the NCBI for that genomic product (usually an EST).

Value

A character vector containing the query string.

Note

Be very careful about automatically querying this resource. It is considered antisocial behavior by the owners.

Author(s)

Robert Gentleman

References

NCBI, https://www.ncbi.nih.gov/

Examples


  q1<-UniGeneQuery(c("Hs.293970", "Hs.155650"))
  q1
  if( interactive())
    browseURL(q1[1])


[Package annotate version 1.60.1 Index]