annPkgName {annotate}	R Documentation

Get annotation package name from chip name

Description

This function returns the name of the Bioconductor annotation data package that corresponds to the specified chip or genome. The type argument is used to request an annotation package with a particular backing store.

Usage


annPkgName(name, type = c("db", "env"))

Arguments

name
string specifying the name of the chip or genome. For example, "hgu133plus2"

type
Either "db" or "env". This will determine whether the package name returned corresponds to the SQLite-based annotation package or environment-based package, respectively.

Value

a string giving the name of the annotation data package

Author(s)

Seth Falcon

See Also

getAnnMap

Examples


annPkgName("hgu133plus2", type="db")
annPkgName("hgu133plus2", type="env")

[Package annotate version 1.60.1 Index]