buildChromLocation {annotate}	R Documentation

A function to generate an instantiation of a chromLocation class

Description

This function will take the name of a data package and build a chromLocation object representing that data set.

Usage


  buildChromLocation(dataPkg)

Arguments

dataPkg
The name of the data package to be used

Details

The requested data set must be available in the user's .libPaths(), and the function will throw an error if this is not the case.

If the data package is present, the necessary information will be extracted from the data package and a chromLocation object will be created.

Value

A chromLocation object representing the specified data set.

Author(s)

Jeff Gentry

Examples


  library("hgu95av2.db")
  z <- buildChromLocation("hgu95av2")

[Package annotate version 1.60.1 Index]