buildPubMedAbst {annotate}	R Documentation

A function to generate an instantiation of a pubMedAbst class

Description

This function will take in a XML tree object and will create an instance of a pubMedAbst class. This instance is returned to the caller.

Usage


  buildPubMedAbst(xml)

Arguments

xml
A XMLTree object that corresponds to a Pubmed abstract.

Value

This function returns an instantiation of a pubMedAbst object to the caller.

Author(s)

Jeff Gentry

See Also

pubmed,genbank

Examples


   x <- pubmed("9695952","8325638","8422497")
   a <- xmlRoot(x)
   numAbst <- length(xmlChildren(a))
   absts <- list()
   for (i in 1:numAbst) {
      absts[[i]] <- buildPubMedAbst(a[[i]])
   }


[Package annotate version 1.60.1 Index]