chromLocation-class {annotate}	R Documentation

Class chromLocation, a class for describing genes and their chromosome mappings.

Description

This class provides chromosomal information provided by a Bioconductor metadata package. By creating the object once for a particular package, it can be used in a variety of locations without the need to recomputed values repeatedly.

Creating Objects

new('chromLocation', organism = ...., # Object of class character
dataSource = ...., # Object of class character
chromLocs = ...., # Object of class list
probesToChrom = ...., # Object of class ANY
chromInfo = ...., # Object of class numeric
geneSymbols = ...., # Object of class ANY
)

Slots

organism:

Object of class "character". The organism that these genes correspond to.

dataSource:

Object of class "character". The source of the gene data.

chromLocs:

Object of class "list". A list which provides specific location information for every gene.

probesToChrom:

An object with an environment-like API which will translate a probe identifier to chromosome it belongs to.

chromInfo:

A numerical vector representing each chromosome, where the names are the names of the chromosomes and the values are their lengths

geneSymbols:

An environment or an object with environment-like API that maps a probe ID to the appropriate gene symbol

Methods

chromLengths

(chromLocation): Gets the lengths of the chromosome for this organism

chromLocs

(chromLocation): Gets the 'chromLocs' attribute.

chromNames

(chromLocation): Gets the name of the chromosomes for this organism

dataSource

(chromLocation): Gets the 'dataSource' attribute.

probesToChrom

(chromLocation): Gets the 'probesToChrom' attribute.

nChrom

(chromLocation): gets the number of chromosomes this organism has

organism

(chromLocation): gets the 'organism' attribute.

chromInfo

Gets the 'chromInfo' attribute.

geneSymbols

Gets the 'geneSymbols' attribute.

See Also

buildChromLocation

Examples


  library("hgu95av2.db")

  z <- buildChromLocation("hgu95av2")
  
  ## find the number of chromosomes
  nChrom(z)

  ## Find the names of the chromosomes
  chromNames(z)

  ## get the organism this object refers to
  organism(z)

  ## get the lengths of the chromosomes in this object
  chromLengths(z)

[Package annotate version 1.60.1 Index]