compatibleVersions {annotate}	R Documentation

function to check to see if the packages represented by the names passed have the same version number

Description

This function takes the names of installed R packages and then checks to see if they all have the same version number.

Usage


compatibleVersions(...)

Arguments

...
... character strings for the names of R packages that have been installed

Details

If all the package have the same version number, the function returns TRUE. Otherwise, the function returns FALSE

Value

This function returns TRUE or FALSE depending on whether the packages have the same version number

Author(s)

Jianhua Zhang

See Also

packageDescription

Examples


  library("hgu95av2.db")
  library("GO.db")
  compatibleVersions("hgu95av2.db", "GO.db")

[Package annotate version 1.60.1 Index]