dropECode {annotate}	R Documentation

Drop GO labels for specified Evidence Codes

Description

Genes are mapped to GO terms on the basis of evidence codes. In some analyses it will be appropriate to drop certain sets of annotations based on specific evidence codes.

Usage


dropECode(inlist, code="IEA")

Arguments

inlist
A list of GO data

code
The set of codes that should be dropped.

Details

A simple use of lapply and sapply to find and eliminate those terms that have the specified evidence codes.

This might be used when one is using to GO to validate a sequence matching experiment (for example), then all terms whose mapping was based on sequence similarity (say ISS and IEA) should be removed.

Value

A list of the same length as the input list retaining only those annotations whose evidence codes were not the ones in the exclusion set code.

Author(s)

R. Gentleman

See Also

getEvidence, getOntology

Examples


 library("hgu95av2.db")
 bb <- hgu95av2GO[["39613_at"]]
 getEvidence(bb[1:3])
 cc <- dropECode(bb[1:3])
 if (length(cc))
   getEvidence(cc)

[Package annotate version 1.60.1 Index]