entrezGeneByID {annotate}	R Documentation

Create a Query String for an Entrez Gene Identifier

Description

Given a set of UniGene identifiers this function creates a set of URLs that an be used to either open a browser to the requested location or that can be used as anchors in the construction of HTML output.

Usage


entrezGeneByID(query)

Arguments

query
Entrez Gene identifiers.

Details

Using NCBI we construct appropriate strings for directing a web browser to the Entrez Genes specified by their IDs.

Value

A character vector containing the query string.

Note

Be very careful about automatically querying this resource. It is considered antisocial behavior by the owners.

Author(s)

Marc Carlson

References

NCBI, https://www.ncbi.nih.gov/

Examples


  q1<-entrezGeneByID(c("100", "1002"))
  q1
  if( interactive())
    browseURL(q1[1])


[Package annotate version 1.60.1 Index]