entrezGeneQuery {annotate}	R Documentation

Create a Query String for Entrez Genes

Description

Given a set of search terms this function creates a set of URLs that an be used to either open a browser to the requested location or that can be used as anchors in the construction of HTML output.

Usage


entrezGeneQuery(query)

Arguments

query
The UniGene identifiers.

Details

Using NCBI we construct an appropriate string for directing a web browser to information about genes of that type at NCBI.

Value

A character vector containing the query string.

Note

Be very careful about automatically querying this resource. It is considered antisocial behavior by the owners.

Author(s)

Marc Carlson

References

NCBI, https://www.ncbi.nih.gov/

Examples


  q1<-entrezGeneQuery(c("leukemia", "Homo sapiens"))
  q1
  if( interactive())
    browseURL(q1[1])


[Package annotate version 1.60.1 Index]