getAnnMap {annotate}	R Documentation

Get annotation map

Description

This function retrieves a map object from an annotation data package. It is intended to serve as a common interface for obtaining map objects from both SQLite-based and environment-based annotation data packages.

Usage


getAnnMap(map, chip, load = TRUE, type = c("db", "env"))

Arguments

map
a string specifying the name of the map to retrieve. For example, "ENTREZID" or "GO"

chip
a string describing the chip or genome

load
a logical value. When TRUE, getAnnMap will try to load the annotation data package if it is not already attached.

type
a character vector of one or more annotation data package types. The currently supported types are "db" and "env". If load is TRUE, you can specify both "db" and "env" and the order will determine which type is tried first. This provides a fall-back mechanism when the preferred annotation data package type is not available. If type is missing, then the first matching annotation package found in the search path will be used, and then the default value of type takes over.

Details

getAnnMap uses the search path (see search) to find an appropriate annotation data package; when called with chip="hgu95av2", the function will use the first hgu95av2 package on the search path whether it be db or environment-based. If load=TRUE and no suitable package is found on the search path, then the function will attempt to load an appropriate package. The type argument is used to determine which type of package (db or env) is loaded first.

Value

If type is "db", an S4 object representing the requested map. If type is "env", an R environment object representing the requested map.

Author(s)

Seth Falcon

Examples


map <- getAnnMap("ENTREZID", "hgu95av2", load=TRUE, type=c("env", "db"))
class(map)

[Package annotate version 1.60.1 Index]