getEvidence {annotate}	R Documentation

Get the Evidence codes for a set of GO terms.

Description

For each mapping of a gene to a GO term there are a set of evidence codes that are used. Genes can be mapped using one, or more evidence codes and this function obtains the evidence codes for all genes provided in the input list.

Usage


getEvidence(inlist)

Arguments

inlist
A list of GO identifers.

Value

A list of the same length as the input list, each element is a vector of evidence codes.

Author(s)

R. Gentleman

See Also

getOntology, dropECode

Examples


 library("hgu95av2.db")
 bb <- hgu95av2GO[["39613_at"]]
 getEvidence(bb)

[Package annotate version 1.60.1 Index]