getOntology {annotate}	R Documentation

Get GO terms for a specified ontology

Description

Find the subset of GO terms for the specified ontology, for each element of the supplied list of associations. The input list is typically from one of the chip-specific meta-data files.

Usage


getOntology(inlist, ontology=c("MF", "BP", "CC"))

Arguments

inlist
A list of GO associations

ontology
The name of the ontology you want returned.

Details

The input list should be a list of lists, each element of inlist is itself a list containing the information that maps from a specified ID (usually LocusLink) to GO information. Each element of the inner list is a list with elements GOID, Ontology and Evidence.

Value

A list of the same length as the input list. Each element of this list will contain a vector of GOIDs for those terms that match the requested ontology.

Author(s)

R. Gentleman

See Also

getEvidence, dropECode

Examples


 library("hgu95av2.db")
 bb <- hgu95av2GO[["39613_at"]]
 getOntology(bb)
 sapply(bb, function(x) x$Ontology)

[Package annotate version 1.60.1 Index]