getPMInfo {annotate}	R Documentation

extract publication details and abstract from annotate::pubmed function output

Description

extract publication details and abstract from annotate::pubmed function output

Usage


getPMInfo(x)

Arguments

x
an object of class xmlDocument; assumed to be result of a pubmed() call

Details

uses xmlDOMApply to extract and structure key features of the XML tree returned by annotate::pubmed()

Value

a list with one element per pubmed id processed by pubmed. Each element of the list is in turn a list with elements for author list, title, journal info, and abstract text.

Note

this should be turned into a method returning an instance of a formal class representing articles.

Author(s)

Vince Carey <stvjc@channing.harvard.edu>

Examples


demo <- pubmed("11780146", 
    "11886385", "11884611")
getPMInfo(demo)

[Package annotate version 1.60.1 Index]