getSEQ {annotate}	R Documentation

Queries the NCBI database to obtain the sequence for a given GenBank Accession number

Description

Given a GenBank Accession number, getSEQ queries the NCBI database for the nucleotide sequence.

Usage


getGI(accNum)
getSEQ(gi)

Arguments

accNum
accNum a character string for a GenBank Accession number (i.e. M22490)

gi
gi a character string or numeric numbers for a GenBank accession number or gi number. A gi number is a series of digits that are assigned consecutively to each sequence record processed by NCBI

Details

The NCBI database is queried for the given GenBank Accession number to obtain the nucleotide sequence in FASTA format. The leading identification line of the sequence data is then dropped to return only the nucleotide sequence.

getGI returns the gi number corresponding to a given GenBank accession number.

Value

getSEQ returns a character string of nucleotide sequence

Author(s)

Jianhua Zhang

References

https://www.ncbi.nlm.nih.gov/entrez/query.fcgi

Examples


getSEQ("M22490")

[Package annotate version 1.60.1 Index]