hasGOannote {annotate}	R Documentation

Check for GO annotation

Description

Given a GO term, or a vector of GO terms and an ontology this function determines which of the terms have GO annotation in the specified ontology.

Usage


hasGOannote(x, which="MF")

Arguments

x
A character vector, an instance of the GOTerms class or a list of GOTerms.

which
One of "MF", "BP" or "CC"

Details

The available GO annotation is searched and a determination of whether a specific GO identifier has a value in the specified ontology is made.

Value

A logical vector of the same length as x.

Author(s)

R. Gentleman

See Also

get

Examples


 library("GO.db")
 t1 <- "GO:0003680"
 hasGOannote(t1)
 hasGOannote(t1, "BP")

[Package annotate version 1.60.1 Index]