hgByChroms {annotate}	R Documentation

A dataset to show the human genome base pair locations per chromosome.

Description

The data is described above.

Usage

data(hgByChroms)

Format

A list, with the names consisting of the names of the chromosomes in the human genome (thus 24 elements). Each element consists of a named vector of +/- values - where each value represents the location of a base pair (the numeric value is the location, while the +/- denotes the strand value), with the name providing the name of the base pair.

Source

Cheng Li of the Dana-Farber Cancer Institute.

Examples


 data(hgByChroms)

[Package annotate version 1.60.1 Index]