hgu95Asym {annotate}	R Documentation

Annotation data for the Affymetrix HGU95A GeneChip

Description

Data, in the form of environments for the Affymetrix U95A chip.

Usage

data(hgu95Asym)

Format

This data set provides an environment with mappings from the Affymetrix identifiers to gene symbol. The environment functions like a hashtables and can be accessed using mget. If the returned value is NA then the current mapping was unable to identify this. Mappings and data sources are constantly evolving so updating often is recommended.

Source

The AnnBuilder package.

Examples


 data(hgu95Asym)
 data(sample.ExpressionSet)
 mget(featureNames(sample.ExpressionSet)[330:340], env=hgu95Asym, ifnotfound=NA)

[Package annotate version 1.60.1 Index]