mapOrgs {annotate}	R Documentation

Functions to map to organism IDs used by NCBI homology.

Description

These functions help map to organism identifiers used at the NCBI.

Usage


mapOrgs(toMap, what = c("code","name"))
getOrgNameNCode()

Arguments

toMap
vect a vector of character strings

what
what a character string that can either be "code" or "name".

Details

mapOrgs converts organism codes to scientific names.

Value

mapOrgs returns a vector of character strings.

Author(s)

Jianhua Zhang

References

ftp://ftp.ncbi.nih.gov/pub/HomoloGene/README

[Package annotate version 1.60.1 Index]