organism {annotate}	R Documentation

Convenience function for getting the organism from an object or package

Description

The most basic organism method just takes a character string (which represents a particular annotation package) and returns the organism that said package is based upon.

Usage


  organism(object)

Arguments

object
a character string that names a package

Value

The name of the organism used for this package or object

Author(s)

Marc Carlson

Examples


  require(hgu95av2.db)
  ## get the organism for this annotation package
  organism("hgu95av2")

  ## get the organism this object refers to
  ## (for a ChromLocation object)
  z <- buildChromLocation("hgu95av2")
  organism(z)


[Package annotate version 1.60.1 Index]