p2LL {annotate}	R Documentation

A function to map from probes to unique Entrez Gene IDs

Description

For any chip, this function computes the map from unique Entrez Gene ID to all probes.

Usage


p2LL(data)

Arguments

data
The character string naming the chip.

Details

This function is deprecated.

This is essentially the computation of the reverse map, we store probe to Entrez gene information in the ENTREZID environment. This is used to compute the inverse mapping.

Value

A list, with length equal to the number of unique Entrez Gene IDs on the chip, the elements correspond to the probes that map to the Gene ID.

Author(s)

R. Gentleman

See Also

getEG

Examples


## Not run: 
  library("hgu95av2.db")
  x <- p2LL("hgu95av2")
  table(sapply(x, length))

## End(Not run)

[Package annotate version 1.60.1 Index]