pm.abstGrep {annotate}	R Documentation

An interface to grep for PubMed abstracts.

Description

A user friendly interface to the functionality provided by pubmed.

Usage


pm.abstGrep(pattern, absts, ...)

Arguments

pattern
A pattern for the call to grep.

absts
A list containing abstracts downloaded using pubmed or equivalent.

...
Extra arguments passed to grep.

Details

The absts are a list of PubMed XML objects that have been downloaded and parsed. This function lets the user quickly search the abstracts for any regular expression. The returned value is a logical vector indicating which of the abstracts contain the regular expression.

Value

The returned value is a logical vector indicating which of the abstracts contain the regular expression.

Author(s)

Robert Gentleman

See Also

pm.getabst, pm.titles

Examples


  library("hgu95av2.db")
  hoxa9 <- "37806_at"
  absts <- pm.getabst(hoxa9, "hgu95av2")
  pm.abstGrep("SH3", absts[[1]])
  pm.abstGrep("autism", absts[[1]])

[Package annotate version 1.60.1 Index]