pm.titles {annotate}	R Documentation

Obtain the titles of the PubMed abstracts.

Description

This function returns the titles from a list of PubMed abstracts.

Usage


pm.titles(absts)

Arguments

absts
The list of PubMed abstracts.

Details

It simply uses sapply.

Value

A character vector of length equal to the number of abstracts. Each element is the title of the corresponding abstract.

Author(s)

Robert Gentleman

See Also

pm.abstGrep

Examples


  library("hgu95av2.db")
  hoxa9 <- "37806_at"
  absts <- pm.getabst(hoxa9, "hgu95av2")
  pm.titles(absts)[[1]][[1]]

[Package annotate version 1.60.1 Index]