pmidQuery {annotate}	R Documentation

A function to query PubMed

Description

Given a PMID, will create a URL which can be used to open a browser and retrieve the specified information from PubMed.

Usage


pmidQuery(query)

Arguments

query
The PubMed ID (or IDs)

Details

Using ublished details from NCBI we construct an appropriate string for directing a web browser to the information available at the NCBI.

Value

A character string containing the appropriate URL

Author(s)

Jeff Gentry

References

NCBI, https://www.ncbi.nih.gov/

See Also

UniGeneQuery

Examples


  a <- "9695952"
  pmidQuery(a)

[Package annotate version 1.60.1 Index]