pubmed {annotate}	R Documentation

A function to open the browser to Pubmed with the selected gene.

Description

Given a vector of Pubmed identifiers or accession numbers, the user can either have a browser display a URL showing a Pubmed query for those identifiers, or a XMLdoc object with the same data.

Usage


pubmed(...,disp=c("data","browser"), type=c("uid","accession"),
       pmaddress=.efetch("PubMed", disp, type))

Arguments

...
Vectorized set of Pubmed ID's

disp
Either "Data" or "Browser" (default is data). Data returns a XMLDoc, while Browser will display information in the user's browser.

type
Denotes whether the arguments are accession numbers or UIDS. Defaults to uids.

pmaddress
Specific path to the pubmed efetch engine from the NCBI website.

Details

A simple function to retrieve Pubmed data given a specific ID, either through XML or through a web browser. This function will accept either pubmed accession numbers or NCBI UIDs (defined as a Pubmed ID or a Medline ID) - although the types must not be mixed in a single call.

WARNING: The powers that be at NCBI have been known to ban the IP addresses of users who abuse their servers (currently defined as less then 2 seconds between queries). Do NOT put this function in a tight loop or you may find your access revoked.

Value

If the option "data" is used, an object of type XMLDoc is returned, unless there was an error with the query in which case an object of type try-error is returned.

If the option "browser" is used, nothing is returned.

Author(s)

R. Gentleman

See Also

genbank, xmlTreeParse

Examples


   if( interactive() )
    opts <- c("data","browser") else
    opts <- "data"
   for (dp in opts)
     pubmed("11780146","11886385","11884611",disp=dp)

[Package annotate version 1.60.1 Index]