serializeEnv {annotate}	R Documentation

A Function To Serialize Environment

Description

This function will serialize an environment in R to an XML format stored in a compressed file.

Usage


serializeEnv(env, fname)
serializeDataPkgEnvs(pkgDir)

Arguments

env
The name of the environment to serialize.

fname
The name of the output file.

pkgDir
The directory where a data package is

Details

The environment is converted into an XML format and then outputted to a gzipped file (using gzfile). The values in the environment are serialized (using serialize) in ASCII format although the keys are stored in plain text.

The format of the XML is very simple, with the primary block being values, which contain blocks of entries, and each entry having a key and a value. For instance, if we had an environment with one value in it, the character c with a key of a (e.g. assign("a", "c", env=foo)), this is what the output would look like.


     <?xml version="1.0"?>
     <values xmlns:bt="http://www.bioconductor.org/RGDBM">
        <entry>
           <key>
             a
           </key>
           <value>
              A\n2\n131072\n66560\n1040\n1\n1033\n1\nc\n
           </value>
	</entry>    
     </values>
   

Author(s)

Jeff Gentry

See Also

gzfile, serialize

Examples


   z <- new.env()
   assign("a", 1, env=z)
   assign("b", 2, env=z)
   assign("c", 3, env=z)
   serializeEnv(z, tempfile())

[Package annotate version 1.60.1 Index]