usedChromGenes {annotate}	R Documentation

A function to select used genes on a chromosome from an ExpressionSet.

Description

Given an instance of an ExpressionSet, a chromLocation object and the name of a chromosome this function returns all genes represented in the ExpressionSet on the specified chromosome.

Usage


usedChromGenes(eSet, chrom, specChrom)

Arguments

eSet
An instance of an ExpressionSet object.

chrom
The name of the chromosome of interest.

specChrom
An instance of a chromLocation object.

Value

Returns a vector of gene names that represent the genes from the ExpressionSet that are on the specified chromosome.

Author(s)

Jeff Gentry

Examples


    data(sample.ExpressionSet)
    data(hgu95AProbLocs)
    usedChromGenes(sample.ExpressionSet, "1", hgu95AProbLocs)

[Package annotate version 1.60.1 Index]